replace user select, update with filter
